src/marketplace_backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --allow-unsafe --output-file=src/marketplace_backend/requirements.txt --strip-extras src/marketplace_backend/requirements.in
#
ic-basilisk==0.11.25
    # via
    #   -r src/marketplace_backend/requirements.in
    #   ic-basilisk-toolkit
ic-basilisk-toolkit==0.1.5
    # via -r src/marketplace_backend/requirements.in
ic-python-db==0.7.9
    # via -r src/marketplace_backend/requirements.in
ic-python-logging==0.3.4
    # via -r src/marketplace_backend/requirements.in
